+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Added:
 
 - Niche cluster connectivity visualization for each sample
 - `scale factor` controlling the size of spatial-based plots
+- `N_GCN_LAYERS` controlling the number of GCN layers in the model
 
 Changed: